Switch to flake8 linter
The flake8 linter is a wrapper around pep8 and pyflakes. In addition to
checking for PEP style compliance, flake8 will also find bugs with an
essentially zero false positive rate. As such, it's well worth running
the flake8 linter instead of pep8. A full listing of the issues that
flake will detect can be found here:
http://flake8.pycqa.org/en/latest/user/error-codes.html